requirements.txt
openai==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
numpy<1.25
pandas==2.0.3
python-dotenv==1.0.0
scipy==1.11.1 